fix(knowledge-gate): close cross-directive denial leak and unwired reset gap
PR review of #1893 independently confirmed (with executed probe tests) two
real bugs in the deadlock-escape-hatch fix:

1. The per-session denial counter was keyed only by sessionID, not by which
   critical-directive set it was accrued against. When the injector swaps in
   an unrelated directive on an ordinary phase/task transition
   (setCriticalShownIds), the new directive silently inherited the old one's
   denial count and could get auto-cleared after far fewer than
   max_gate_denials real denials against it.
2. The new denial-count map lived as a private module-scoped variable in
   knowledge-application-gate.ts and was never wired into resetSwarmState /
   resetSwarmStatePreservingSingletons, unlike its sibling maps
   (currentCriticalShownIds, knowledgeAckDedup) -- a stale count could
   survive a same-process /swarm close and leak into a reused session.

Fix: moved the counter into swarmState.gateDenialCounts (src/state.ts),
matching the existing currentCriticalShownIds/knowledgeAckDedup pattern
exactly -- incrementGateDenialCount/clearGateDenialCount now key each entry
to a directiveKey fingerprint (buildGateDenialDirectiveKey, a sorted join of
the directive-id set) so re-injecting the SAME directive across turns
doesn't reset the count, but swapping to a genuinely different directive
does. resetSwarmState() now clears gateDenialCounts alongside its siblings,
closing the session-teardown leak.

Also, per review: the escape-hatch audit-event writes are now awaited
before the gate returns (were fire-and-forget, contradicting the release
note's "never silent" claim); the escape hatches now clear
currentCriticalShownIds via the existing centralized clearCriticalShownIds()
helper instead of a direct Map.delete(); corrected an overclaim in the
release note about skill_regenerate/skill_retire being gated under the
production default config (they aren't, due to a pre-existing, out-of-scope
shadowing bug); and updated docs/knowledge.md + the architect system prompt
to describe the bounded escape hatches instead of claiming unconditional
blocking.

Split two test files that grew past the repo's 500-line limit (AGENTS.md
invariant 7) into focused files, and added regression tests for both bugs
plus FIFO-eviction coverage for the new gateDenialCounts map.
